Move CallEventContent to the separate file
